requirements.txt
flask
flask_restful
camelot-py[cv]
json_tools
json_delta
requests
gunicorn